requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlmodel==0.0.22
psycopg2-binary==2.9.11
python-dotenv==1.0.1